// features/api/model/editRequestsIndex200.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { EditRequestResource } from './editRequestResource';
import type { EditRequestsIndex200Links } from './editRequestsIndex200Links';
import type { EditRequestsIndex200Meta } from './editRequestsIndex200Meta';

export type EditRequestsIndex200 = {
  data: EditRequestResource[];
  links: EditRequestsIndex200Links;
  meta: EditRequestsIndex200Meta;
};