// features/api/model/updateEditRequestRequest.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { UpdateEditRequestRequestStatus } from './updateEditRequestRequestStatus';

export interface UpdateEditRequestRequest {
  status: UpdateEditRequestRequestStatus;
  /** @nullable */
  comments?: string | null;
}